use the float property of the server's timerefs
git-svn-id: svn+ssh://rcs.lan-puteaux.avencall.com/var/lib/svn/xivo/trunk/cticlients@5188 dc85baef-7018-0410-bfbd-8a639da648bb

+++ common/agentdetailspanel.cpp
@@ -136,7 +136,7 @@ void AgentdetailsPanel::setUserInfo(const UserInfo * ui)
         m_userinfo = ui;
 }
 
-void AgentdetailsPanel::updatePeerAgent(int,
+void AgentdetailsPanel::updatePeerAgent(double,
                                         const QString &,
                                         const QString & what,
                                         const QVariant & params)
